requirements.txt
python-dotenv==0.21.0
requests==2.28.1
schedule==1.1.0
tweepy==4.10.1
reverse_geocoder==1.5.1
